JSP에서 타임리프로 변경

+++ injucksung/src/main/java/com/injucksung/injucksung/service/UserServiceImpl.java
@@ -1,24 +1,38 @@
 package com.injucksung.injucksung.service;
 
+import com.injucksung.injucksung.domain.Role;
 import com.injucksung.injucksung.domain.User;
+import com.injucksung.injucksung.repository.RoleRepository;
 import com.injucksung.injucksung.repository.UserRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.HashSet;
+import java.util.Set;
 
 @Service
 public class UserServiceImpl implements UserService {
 
     private UserRepository userRepository;
+    private RoleRepository roleRepository;
+    private static final int PAGE_SIZE = 10;
 
-    public UserServiceImpl(UserRepository userRepository) {
+    public UserServiceImpl(UserRepository userRepository, RoleRepository roleRepository) {
         this.userRepository = userRepository;
+        this.roleRepository = roleRepository;
     }
 
     @Override
     @Transactional
     public int signup(User user) {
+        // TODO: 18. 12. 17 보라색 밑줄이 매우 거슬리니 나중에 modify와 합칠지 고민해 보세
         User signupUser = userRepository.save(user);
         if (signupUser != null) {
+            Set<Role> roles = new HashSet<>();
+            roles.add(roleRepository.getOne(2L));
+            signupUser.setRoles(roles);
             userRepository.flush();
             return 1;
         }
@@ -54,13 +68,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public int deleteAccount(String email, String password) {
+    public void deleteAccount(String email, String password) {
         User user = userRepository.findUserByEmailAndPassword(email, password);
-        if (user != null) {
-            userRepository.delete(user);
-            return 1;
-        }
-        return 0;
+        if (user != null) userRepository.delete(user);
     }
 
     @Override
@@ -73,5 +83,47 @@ public class UserServiceImpl implements UserService {
         }
 
         return 0;
+    }
+
+    @Override
+    @Transactional
+    public int modifyUserRole(Long id, Long[] roleIds) {
+        User user = userRepository.getOne(id);
+        Set<Role> roles = user.getRoles();
+        roles.clear();
+        if (roles.isEmpty()) {
+            for (Long roleId : roleIds) {
+                Role role = roleRepository.getOne(roleId);
+                roles.add(role);
+            }
+        }
+        user.setRoles(roles);
+        if (userRepository.save(user) != null) {
+            userRepository.flush();
+            return 1;
+        }
+        return 0;
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public Page<User> getUserList(int start, String searchType, String searchWord) {
+        PageRequest pageRequest = PageRequest.of(start, PAGE_SIZE);
+        Page<User> users = null;
+        switch (searchType) {
+            case "all":
+                users = userRepository.findAll(pageRequest);
+                break;
+            case "email":
+                users = userRepository.findAllByEmailContaining(searchWord, pageRequest);
+                break;
+            case "nickname":
+                users = userRepository.findAllByNicknameContaining(searchWord, pageRequest);
+                break;
+            case "phoneNo":
+                users = userRepository.findAllByPhoneContaining(searchWord, pageRequest);
+                break;
+        }
+        return users;
     }
 }
